+++ ttnn/core/tensor/tensor_ops.cpp
@@ -23,6 +23,7 @@
 
 #include <tt-metalium/experimental/tensor/tensor_apis.hpp>
 #include <tt-metalium/experimental/distributed_tensor/distributed_tensor_apis.hpp>
+#include <internal/tensor/host_pad_apis.hpp>
 #include <internal/tensor/byte_based_data_transfer_apis.hpp>
 
 using tt::tt_metal::BufferRegion;
